refactor: join type files feat: Introduce Bing maps
