+++ index.html
@@ -3,7 +3,8 @@
 <head>
     <meta charset="UTF-8"/>
     <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-    <meta name="description" content="Dual map - synchronize views of different map types side by side. Compare data from different sources."/>
+    <meta name="description"
+          content="Dual map - synchronize views of different map types side by side. Compare data from different sources."/>
     <title>Dual Map</title>
     <script type="module" src="/src/main.ts"></script>
     <link rel="icon" type="image/svg+xml"
@@ -26,7 +27,7 @@
         <label for="satellite">Satellite</label>
         <input type="radio" id="rail" name="layer-type" value="Rail">
         <label for="rail">Rail</label>
-</nav>
+    </nav>
     <nav class="choice  js-map-type">
         <input type="radio" id="wikimapia" name="map-type" value="wiki">
         <label for="wikimapia">Wikimapia</label>
@@ -34,14 +35,16 @@
         <label for="google-maps">Google maps</label>
         <input type="radio" id="open-railway" name="map-type" value="rail">
         <label for="open-railway">Open railway</label>
+        <input type="radio" id="bing" name="map-type" value="bing">
+        <label for="bing">BING</label>
     </nav>
 </header>
-<div class="layout">
-    <div class="layout__left js-osm">
-    </div>
-    <div class="layout__right js-map">
-    </div>
-</div>
+<main class="layout">
+    <aside class="layout__left js-osm">
+    </aside>
+    <aside class="layout__right js-map">
+    </aside>
+</main>
 <div class="axis axis__horizontal"></div>
 <div class="axis axis__vertical axis__vertical--left"></div>
 <div class="axis axis__vertical axis__vertical--right"></div>
