--- src/Message.type.ts
@@ -1,18 +1,0 @@
-import { MapOptions } from './types.ts';
-
-export enum MessageState {
-  MoveMap = 'MoveMap', //Same as resize
-  KeyPressed = 'KeyPressed',
-}
-
-export type Message =
-  | {
-      state: MessageState.KeyPressed;
-      data: {
-        key: string;
-      };
-    }
-  | {
-      state: MessageState.MoveMap;
-      data: MapOptions;
-    };
