+++ src/Publisher.interface.ts
@@ -1,5 +1,5 @@
-import { type Message } from './Message.type';
 import { Observer } from './Observer.interface.ts';
+import { Message } from './types.ts';
 
 export interface Publisher {
   subscribers: Observer[];
