+++ src/types.ts
@@ -1,4 +1,4 @@
-export type MapType = 'wiki' | 'google' | 'rail' | 'osm';
+export type MapType = 'wiki' | 'google' | 'rail' | 'osm' | 'bing';
 export type LayerName = 'OSM' | 'OSM DE' | 'Humanitarian' | 'Topography' | 'Cyclo' | 'Satellite' | 'Rail';
 
 export type MapOptions = {
@@ -23,6 +23,23 @@ export type UrlParams = {
   type: MapType;
 };
 
+export enum MessageState {
+  MoveMap = 'MoveMap', //Same as resize
+  KeyPressed = 'KeyPressed',
+}
+
+export type Message =
+  | {
+      state: MessageState.KeyPressed;
+      data: {
+        key: string;
+      };
+    }
+  | {
+      state: MessageState.MoveMap;
+      data: MapOptions;
+    };
+
 export type DOMElement = Element | NodeListOf<HTMLElement> | null;
 export type DOMElements = Map<string, DOMElement>;
 export type RadioItemAction = (event: Event) => void;
@@ -30,4 +47,4 @@ export type RadioItemAction = (event: Event) => void;
 export const isLayerName = (value: string): value is LayerName =>
   ['OSM', 'OSM DE', 'Humanitarian', 'Topography', 'Cyclo', 'Satellite', 'Rail'].includes(value);
 
-export const isMapType = (value: string): value is MapType => ['wiki', 'google', 'rail', 'osm'].includes(value);
+export const isMapType = (value: string): value is MapType => ['wiki', 'google', 'rail', 'osm', 'bing'].includes(value);
